remove LangChecks
